Small steps towards making the package
What is life, what is the meaning behind it? Is to live just to suffer? or is there a deeper meaning. Some people say they were born to make a change in someone's life, some say they were born to secure a future for themselves. But in the end, the only meaningful thing I learnt was that dont fkin host half of the world's infrastructure on one server...